fix: 사용자 삭제 시 예외 처리 구체화
- 우선 사용자 존재 여부를 확인 후 삭제하는 로직을 통해 안정성 추가
- 예외 처리 구체화 및 세분화

+++ src/main/java/com/iroff/supportlab/application/user/service/DeleteUserInteractor.java
@@ -1,5 +1,6 @@
 package com.iroff.supportlab.application.user.service;
 
+import org.springframework.dao.DataAccessException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -32,8 +33,12 @@ public class DeleteUserInteractor implements DeleteUserUseCase {
 			.orElseThrow(() -> new DomainException(UserError.USER_NOT_FOUND));
 		checkCondition(passwordEncoder.matches(password, user.getPassword()), AuthError.INVALID_PASSWORD);
 		try {
-			userRepository.deleteById(userId);
-		} catch (Exception e) {
+			if (userRepository.existsById(userId)) {
+				userRepository.deleteById(userId);
+			} else {
+				throw new DomainException(UserError.USER_NOT_FOUND);
+			}
+		} catch (DataAccessException e) {
 			throw new DomainException(UserError.DELETE_USER_FAILED);
 		}
 	}
